Python/FastAPI/swe/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
bcrypt
